add auth decorator; record instupdate datetime

+++ schedulers.py
@@ -1,5 +1,6 @@
 import requests
-from models import db, get_all_institutions, RequestException, ExternalRequestInTransit, TransitStart
+from models import db, get_all_institutions, RequestException, ExternalRequestInTransit, TransitStart, InstUpdate
+from datetime import datetime
 from bs4 import BeautifulSoup
 import sys
 
@@ -8,56 +9,46 @@ def update_reports():
     existing_ext = None  # Initialize variables
     existing_transit = None  # Initialize variables
     institutions = get_all_institutions()  # Get all institutions
+    updated = 0  # Set updated to 0
 
     for institution in institutions:  # loop through institutions
         if institution.code is not None:  # Check if the institution has a code
 
-            ##############
-            # Exceptions #
-            ##############
-
+            # Exceptions
             if institution.exceptions is not None:  # Check if the institution has an exceptions report
-                # Delete existing data for the institution
-                delete_rows(RequestException, institution.code)
-
-                # Get updated report for the institution
-                updated_exc = api_call('exceptions', institution)  # Make the API call
-
-                # map columns to database columns
-                mapped_exc = map_columns(updated_exc)  # dictionary to map column names to column numbers
-
+                delete_rows(RequestException, institution.code)  # Delete existing data for the institution
+                updated_exc = api_call('exceptions', institution)  # Get updated report for the institution
+                mapped_exc = map_columns(updated_exc)  # map columns to database columns
                 exceptions = updated_exc.find_all('Row')  # Get rows from the BeautifulSoup object
+                if exceptions is not None:  # Check if there are any exceptions
+                    for row in exceptions:  # loop through rows
+                        exception = RequestException(  # create a RequestException object
+                            instcode=institution.code,
+                            borreqstat=set_value(row, mapped_exc, 'Borrowing Request Status'),
+                            internalid=set_value(row, mapped_exc, 'Internal Id'),
+                            borcreate=set_value(row, mapped_exc, 'Borrowing Creation Date'),
+                            title=set_value(row, mapped_exc, 'Title'),
+                            author=set_value(row, mapped_exc, 'Author'),
+                            networknum=set_value(row, mapped_exc, 'Network Number'),
+                            isbn=set_value(row, mapped_exc, 'ISBN (Normalized)'),
+                            issn=set_value(row, mapped_exc, 'ISSN (Normalized)'),
+                            requestor=set_value(row, mapped_exc, 'Full Name'),
+                            partnerstat=set_value(row, mapped_exc, 'Partner Active Status'),
+                            reqsend=set_value(row, mapped_exc, 'Request Sending Date'),
+                            days=set_value(row, mapped_exc,
+                                           ' TIMESTAMPDIFF( SQL_TSI_DAY , Request Sending Date,  CURRENT_DATE )'),
+                            partnername=set_value(row, mapped_exc, 'Partner Name'),
+                            partnercode=set_value(row, mapped_exc, 'Partner Code'),
+                        )
+                        try:
+                            db.session.add(exception)  # Add the exception to the database
+                        except Exception as e:
+                            print(e, file=sys.stderr)
+                        else:
+                            db.session.commit()  # Commit the exception to the database
+                    updated += 1  # Increment updated
 
-                for row in exceptions:  # loop through rows
-                    exception = RequestException(  # create a RequestException object
-                        instcode=institution.code,
-                        borreqstat=set_value(row, mapped_exc, 'Borrowing Request Status'),
-                        internalid=set_value(row, mapped_exc, 'Internal Id'),
-                        borcreate=set_value(row, mapped_exc, 'Borrowing Creation Date'),
-                        title=set_value(row, mapped_exc, 'Title'),
-                        author=set_value(row, mapped_exc, 'Author'),
-                        networknum=set_value(row, mapped_exc, 'Network Number'),
-                        isbn=set_value(row, mapped_exc, 'ISBN (Normalized)'),
-                        issn=set_value(row, mapped_exc, 'ISSN (Normalized)'),
-                        requestor=set_value(row, mapped_exc, 'Full Name'),
-                        partnerstat=set_value(row, mapped_exc, 'Partner Active Status'),
-                        reqsend=set_value(row, mapped_exc, 'Request Sending Date'),
-                        days=set_value(row, mapped_exc,
-                                       ' TIMESTAMPDIFF( SQL_TSI_DAY , Request Sending Date,  CURRENT_DATE )'),
-                        partnername=set_value(row, mapped_exc, 'Partner Name'),
-                        partnercode=set_value(row, mapped_exc, 'Partner Code'),
-                    )
-                    try:
-                        db.session.add(exception)  # Add the exception to the database
-                    except Exception as e:
-                        print(e, file=sys.stderr)
-                    else:
-                        db.session.commit()  # Commit the exception to the database
-
-            ################################
-            # External Requests in Transit #
-            ################################
-
+            # External Requests in Transit
             if institution.ext_requests_in_transit is not None:
                 # Get existing data for the institution
                 try:
@@ -70,55 +61,53 @@ def update_reports():
                 mapped_ext = map_columns(updated_ext)  # Map columns to database columns
                 ext_requests = updated_ext.find_all('Row')  # Get rows from the BeautifulSoup object
                 updated_req_ids = []  # Create empty list to hold updated request ids
+                if ext_requests is not None:  # Check if there are any external requests in transit
+                    for request in ext_requests:  # loop through rows
+                        ext_req = ExternalRequestInTransit(  # Create a RequestException object
+                            request_id=set_value(request, mapped_ext, 'Request Id'),
+                            instcode=institution.code,
+                            external_id=set_value(request, mapped_ext, 'External Id'),
+                            requestor=set_value(request, mapped_ext, 'Full Name'),
+                            title=set_value(request, mapped_ext, 'Title'),
+                            author=set_value(request, mapped_ext, 'Author'),
+                            barcode=set_value(request, mapped_ext, 'Barcode'),
+                            physical_item_id=set_value(request, mapped_ext, 'Physical Item Id'),
+                            isbn=set_value(request, mapped_ext, 'ISBN (Normalized)'),
+                            issn=set_value(request, mapped_ext, 'ISSN (Normalized)'),
+                        )
+                        updated_req_ids.append(ext_req.request_id)  # Add the request id to the list
 
-                for request in ext_requests:  # loop through rows
-                    ext_req = ExternalRequestInTransit(  # Create a RequestException object
-                        request_id=set_value(request, mapped_ext, 'Request Id'),
-                        instcode=institution.code,
-                        external_id=set_value(request, mapped_ext, 'External Id'),
-                        requestor=set_value(request, mapped_ext, 'Full Name'),
-                        title=set_value(request, mapped_ext, 'Title'),
-                        author=set_value(request, mapped_ext, 'Author'),
-                        barcode=set_value(request, mapped_ext, 'Barcode'),
-                        physical_item_id=set_value(request, mapped_ext, 'Physical Item Id'),
-                        isbn=set_value(request, mapped_ext, 'ISBN (Normalized)'),
-                        issn=set_value(request, mapped_ext, 'ISSN (Normalized)'),
-                    )
-                    updated_req_ids.append(ext_req.request_id)  # Add the request id to the list
+                        # Check if the updated request id already exists in the database
+                        ext_test = db.session.execute(db.select(ExternalRequestInTransit.request_id)
+                                                      .filter_by(request_id=ext_req.request_id)).scalar_one_or_none()
 
-                    # Check if the updated request id already exists in the database
-                    ext_test = db.session.execute(db.select(ExternalRequestInTransit.request_id)
-                                                  .filter_by(request_id=ext_req.request_id)).scalar_one_or_none()
-
-                    if ext_test is not None:  # If the updated request id already exists in the database
-                        try:
-                            db.session.merge(ext_req)  # Update the external request in transit in the database
-                        except Exception as e:
-                            print(e, file=sys.stderr)
-                            continue
-                        db.session.commit()
-                    else:  # If the updated request id does not exist in the database
-                        try:
-                            db.session.add(ext_req)  # Add the external request in transit to the database
-                        except Exception as e:
-                            print(e, file=sys.stderr)
-                        else:
+                        if ext_test is not None:  # If the updated request id already exists in the database
+                            try:
+                                db.session.merge(ext_req)  # Update the external request in transit in the database
+                            except Exception as e:
+                                print(e, file=sys.stderr)
+                                continue
                             db.session.commit()
+                        else:  # If the updated request id does not exist in the database
+                            try:
+                                db.session.add(ext_req)  # Add the external request in transit to the database
+                            except Exception as e:
+                                print(e, file=sys.stderr)
+                            else:
+                                db.session.commit()
 
-                # Delete old data from the database
-                for request in existing_ext:  # loop through existing external requests in transit
-                    if str(request.request_id) not in updated_req_ids:  # If existing request id not in the updated list
-                        try:
-                            db.session.delete(request)  # Delete the external request in transit from the database
-                        except Exception as e:
-                            print(e, file=sys.stderr)
-                        else:
-                            db.session.commit()
+                    # Delete old data from the database
+                    for request in existing_ext:  # loop through existing external requests in transit
+                        if str(request.request_id) not in updated_req_ids:  # If existing request not in updated list
+                            try:
+                                db.session.delete(request)  # Delete the external request in transit from the database
+                            except Exception as e:
+                                print(e, file=sys.stderr)
+                            else:
+                                db.session.commit()
+                    updated += 1  # Increment updated
 
-            #################
-            # Transit Start #
-            #################
-
+            # Transit Start
             if institution.in_transit_data is not None:
                 # Get existing data for the institution
                 try:
@@ -131,44 +120,54 @@ def update_reports():
                 mapped_transit = map_columns(updated_transit)  # Map columns to database columns
                 transit_data = updated_transit.find_all('Row')  # Get rows from the BeautifulSoup object
                 updated_event_ids = []  # Create empty list to hold updated event ids
+                if transit_data is not None:  # Check if there are any transit events
+                    for event in transit_data:
+                        transit_event = TransitStart(  # Create a TransitStart object
+                            event_id=set_value(event, mapped_transit, 'Event id'),
+                            instcode=institution.code,
+                            request_id=set_value(event, mapped_transit, 'Request id'),
+                            transit_date=set_value(event, mapped_transit, 'Event Start Date and Time'),
+                        )
+                        updated_event_ids.append(transit_event.event_id)  # Add the event id to the list
 
-                for event in transit_data:
-                    transit_event = TransitStart(  # Create a TransitStart object
-                        event_id=set_value(event, mapped_transit, 'Event id'),
-                        instcode=institution.code,
-                        request_id=set_value(event, mapped_transit, 'Request id'),
-                        transit_date=set_value(event, mapped_transit, 'Event Start Date and Time'),
-                    )
-                    updated_event_ids.append(transit_event.event_id)  # Add the event id to the list
+                        # Check if the updated event id already exists in the database
+                        transit_test = db.session.execute(db.select(TransitStart.event_id)
+                                                          .filter_by(
+                            event_id=transit_event.event_id)).scalar_one_or_none()
 
-                    # Check if the updated event id already exists in the database
-                    transit_test = db.session.execute(db.select(TransitStart.event_id)
-                                                      .filter_by(event_id=transit_event.event_id)).scalar_one_or_none()
+                        if transit_test is not None:  # If the updated event id already exists in the database
+                            try:
+                                db.session.merge(transit_event)  # Update the transit start in the database
+                            except Exception as e:
+                                print(e, file=sys.stderr)
+                            else:
+                                db.session.commit()
+                        else:  # If the updated event id does not exist in the database
+                            try:
+                                db.session.add(transit_event)  # Add the transit start to the database
+                            except Exception as e:
+                                print(e, file=sys.stderr)
+                            else:
+                                db.session.commit()
 
-                    if transit_test is not None:  # If the updated event id already exists in the database
-                        try:
-                            db.session.merge(transit_event)  # Update the transit start in the database
-                        except Exception as e:
-                            print(e, file=sys.stderr)
-                        else:
-                            db.session.commit()
-                    else:  # If the updated event id does not exist in the database
-                        try:
-                            db.session.add(transit_event)  # Add the transit start to the database
-                        except Exception as e:
-                            print(e, file=sys.stderr)
-                        else:
-                            db.session.commit()
+                    # Delete old data from the database
+                    for event in existing_transit:  # loop through existing transit starts
+                        if str(event.event_id) not in updated_event_ids:  # If existing event id not in the updated list
+                            try:
+                                db.session.delete(event)  # Delete the transit start from the database
+                            except Exception as e:
+                                print(e, file=sys.stderr)
+                            else:
+                                db.session.commit()
+                    updated += 1  # Increment updated
 
-                # Delete old data from the database
-                for event in existing_transit:  # loop through existing transit starts
-                    if str(event.event_id) not in updated_event_ids:  # If  existing event id  not in the updated list
-                        try:
-                            db.session.delete(event)  # Delete the transit start from the database
-                        except Exception as e:
-                            print(e, file=sys.stderr)
-                        else:
-                            db.session.commit()
+            # Institution Update
+            if updated > 0:
+                inst_update = InstUpdate(institution.code, datetime.now())  # Create an InstUpdate object
+                try:
+                    db.session.add(inst_update)  # add the institution update to the database
+                except Exception as e:
+                    print(e, file=sys.stderr)
 
 
 # Get the value of a column from a row
